+++ src/buildstructor/codegen.rs
@@ -73,7 +73,7 @@ pub fn codegen(ir: Ir) -> Result<TokenStream> {
                 }
             }
 
-           pub struct __Required<T> {
+            pub struct __Required<T> {
                 _phantom: std::marker::PhantomData<T>,
             }
             pub struct __Optional<T> {